requirements.txt
torch
Pillow
matplotlib
torchvision